refactor(extensions)!: move Math extension from yfm to additional folder
