+++ src/bundle/config/wysiwyg.ts
@@ -2,12 +2,12 @@ import {ActionStorage} from 'src/core';
 
 import {headingType, pType} from '../../extensions';
 import {gptHotKeys} from '../../extensions/additional/GPT/constants';
+// for typings from Math
+import type {} from '../../extensions/additional/Math';
 import type {
     SelectionContextConfig,
     SelectionContextItemData,
 } from '../../extensions/behavior/SelectionContext';
-// for typings from Math
-import type {} from '../../extensions/yfm/Math';
 import {i18n as i18nHint} from '../../i18n/hints';
 import {i18n} from '../../i18n/menubar';
 import {Action as A, formatter as f} from '../../shortcuts';
