+++ src/extensions/additional/Math/const.ts
