+++ src/extensions/additional/Math/view-and-edit.scss
